pyproject.toml
[project]
name = "pywebview-base"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "pydantic>=2.12.5",
    "pyinstaller>=6.18.0",
    "pywebview>=6.1",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
